% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{get_available_colours_grid}
\alias{get_available_colours_grid}
\title{get a grid showing the available colours and their indices}
\usage{
get_available_colours_grid(many = T, tons = F)
}
\arguments{
\item{many}{use the 'many' method to get the colours}

\item{tons}{use the 'tons' method to get the colours}
}
\value{
a ggplot grid showing the available colours and their indices
}
\description{
get a grid showing the available colours and their indices
}
\examples{
get_available_colours_grid()
}
