% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{get_color_list}
\alias{get_color_list}
\title{get a vector of as distinct possible colours, but with more possibilities ()}
\usage{
get_color_list(vector_of_names, use_sampling = F, color_indices = NULL)
}
\arguments{
\item{vector_of_names}{a vector of values for which you want colours}

\item{use_sampling}{whether or not to randomly extract the colours instead of grabbing the first n colours}

\item{color_indices}{(optional, not used by default) if specific colours are needed, supply the indices of the colours here. Use 'get_available_colours_grid' to get the colours and their indices}
}
\value{
a vector of colours
}
\description{
get a vector of as distinct possible colours, but with more possibilities ()
}
\examples{
get_color_list(c('A', 'B', 'A', 'C', 'B'), F, c(3,4,5))
get_color_list(c('A', 'B', 'A', 'C', 'B'), T)
}
