% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{get_color_list_shades}
\alias{get_color_list_shades}
\title{get a vector of as distinct possible colours, but with more possibilities ()}
\usage{
get_color_list_shades(vector_of_names, shades_from)
}
\arguments{
\item{vector_of_names}{a vector of values for which you want colours}

\item{shades_from}{vector of colours to compute shades from. Needs to have at least two colours.}
}
\value{
a list of each input value as name, and values of colours
}
\description{
get a vector of as distinct possible colours, but with more possibilities ()
}
\examples{
get_color_list_shades(c('A', 'B', 'A', 'C', 'B'), shades_from = c('lavender', 'darkviolet'))
}
