% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{get_shades}
\alias{get_shades}
\title{get a vector of shades from a supplied vector of colours. Works best if the colours are ordered in a logical order, so yellow, orange, red for example}
\usage{
get_shades(number_of_colours, shades_from)
}
\arguments{
\item{number_of_colours}{the number of colours to get}

\item{shades_from}{vector of colours to compute shades from. Needs to have at least two colours.}
}
\value{
a vector of colours
}
\description{
get a vector of shades from a supplied vector of colours. Works best if the colours are ordered in a logical order, so yellow, orange, red for example
}
\examples{
shades_between_red_yellow <- get_shades(10, c('yellow', 'orange', 'red'))
}
