% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{plot_colours_in_grid}
\alias{plot_colours_in_grid}
\title{get a ggplot2 tile plot with the supplied colours}
\usage{
plot_colours_in_grid(colours_to_plot)
}
\arguments{
\item{colours_to_plot}{a vector of values for which you want colours}
}
\value{
a ggplot2 tile plot with the colours
}
\description{
get a ggplot2 tile plot with the supplied colours
}
\examples{
plot_colours_in_grid(c('darkblue', 'navyblue', 'purple', 'lightblue'))
}
