% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{plot_shades}
\alias{plot_shades}
\title{plot a vector of shades from a supplied vector of colours. Works best if the colours are ordered in a logical order, so yellow, orange, red for example}
\usage{
plot_shades(number_of_colours, shades_from)
}
\arguments{
\item{number_of_colours}{the number of colours to get}

\item{shades_from}{vector of colours to compute shades from. Needs to have at least two colours.}
}
\value{
a ggplot2 tile plot with the colours
}
\description{
plot a vector of shades from a supplied vector of colours. Works best if the colours are ordered in a logical order, so yellow, orange, red for example
}
\examples{
plot_shades(10, c('yellow', 'orange', 'red'))
}
