% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roycols.R
\name{sample_tons_of_colors}
\alias{sample_tons_of_colors}
\title{get a vector of as distinct possible colours, but with more possibilities ()}
\usage{
sample_tons_of_colors(
  number_of_colours,
  use_sampling = F,
  color_indices = NULL
)
}
\arguments{
\item{number_of_colours}{how many colours to return}

\item{use_sampling}{whether or not to randomly extract the colours instead of grabbing the first n colours}

\item{color_indices}{(optional, not used by default) if specific colours are needed, supply the indices of the colours here. Use 'get_available_colours_grid' to get the colours and their indices}
}
\value{
a vector of colours
}
\description{
get a vector of as distinct possible colours, but with more possibilities ()
}
\examples{
colours_to_use <- sample_tons_of_colors(5, F, c(1,4,7,8,11))
colours_to_use <- sample_tons_of_colors(120, T)
}
